separate location domain from bio
